ejbca/requirements/requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
attrs==18.2.0
cached-property==1.5.1
certifi==2018.11.29
cffi==1.10.0
chardet==3.0.4
Click==7.0
colorlog==3.1.4
cryptography==2.5
defusedxml==0.5.0
dojot.module==0.0.1a4
Flask==1.0.2
idna==2.5
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
kafka==1.3.5
kafka-python==1.4.3
lxml==3.7.2
MarkupSafe==1.1.0
packaging==16.8
pycparser==2.17
pyOpenSSL==19.0.0
pyparsing==2.2.0
pytz==2018.9
requests==2.20.0
requests-toolbelt==0.9.1
six==1.10.0
urllib3==1.24.1
Werkzeug==0.14.1
zeep==3.2.0